+++ src/components/alpha/index.js
@@ -1,3 +1,5 @@
 import Alpha from './alpha.vue';
+import { install } from '@/utils/compoent';
 
+Alpha.install = install;
 export default Alpha;