+++ src/components/chrome/index.js
@@ -1,3 +1,5 @@
 import Chrome from './chrome.vue';
+import { install } from '@/utils/compoent';
 
+Chrome.install = install;
 export default Chrome;