+++ src/components/hue/index.js
@@ -1,3 +1,5 @@
 import Hue from './hue.vue';
+import { install } from '@/utils/compoent';
 
+Hue.install = install;
 export default Hue;