+++ src/components/material/index.js
@@ -1,3 +1,5 @@
 import Material from './material.vue';
+import { install } from '@/utils/compoent';
 
+Material.install = install;
 export default Material;