+++ src/components/swatches/index.js
@@ -1,3 +1,5 @@
 import Swatches from './swatches.vue';
+import { install } from '@/utils/compoent';
 
+Swatches.install = install;
 export default Swatches;